Consolidate stdfblib libraries
